moving NCCL helper
