+++ cpp/tensorrt_llm/kernels/userbuffers/ub_allocator.h
@@ -19,11 +19,6 @@
 #if ENABLE_MULTI_DEVICE
 #include "nccl.h"
 #include "userbuffers.h"
-#ifdef _WIN32
-#include <windows.h>
-#else
-#include <dlfcn.h>
-#endif
 #else
 using ncclWindow_t = void*;
 #endif
@@ -80,53 +75,14 @@ protected:
     tensorrt_llm::runtime::WorldConfig mWorldConfig;
 };
 
-class NCCLHelper
-{
-public:
-    NCCLHelper();
-    ~NCCLHelper();
-
-    // Dynamic loading function type definition
-    using ncclCommWindowRegisterFunc = ncclResult_t (*)(ncclComm_t, void*, size_t, ncclWindow_t*, int);
-    using ncclMemAllocFunc = ncclResult_t (*)(void**, size_t);
-
-    // Get function pointer for ncclCommWindowRegister
-    ncclCommWindowRegisterFunc getNCCLCommWindowRegister();
-
-    // Get function pointer for ncclMemAlloc
-    ncclMemAllocFunc getNCCLMemAlloc();
-
-    // Check if NCCL library is successfully loaded
-    bool isLoaded() const;
-
-private:
-    void loadNCCLLibrary();
-    void* loadLibraryHandle(char const* libName);
-    void* getSymbolAddress(void* handle, char const* symbolName);
-
-#ifdef _WIN32
-    HMODULE mLibraryHandle;
-#else
-    void* mLibraryHandle;
-#endif
-
-    ncclCommWindowRegisterFunc mNCCLCommWindowRegister;
-    ncclMemAllocFunc mNCCLMemAlloc;
-    bool mIsLoaded;
-};
-
 class NCCLUserBufferAllocator : public UserBufferAllocator
 {
 public:
     void initialize(tensorrt_llm::runtime::WorldConfig const& world_config) override;
     UBBuffer registerUBBuffer(size_t bytes) override;
 
-    // Get shared NCCLHelper instance
-    static NCCLHelper& getNCCLHelper();
-
 private:
     std::shared_ptr<ncclComm_t> mComm;
-    static std::unique_ptr<NCCLHelper> mNCCLHelper;
 };
 
 #else
